requirements.txt
google-generativeai == 0.3.2
streamlit ==  1.42.0